Run CI with cabal flag +experimental-evaluator
We want to be able to test the experimental evalautor, without
running these tests on branches where the evalautor does not exist.
The CI setup has been modified so any branch matching

  partial-evalautor-*

will build any artifact that uses the partial-evaluator with the
+experimental-evalautor cabal flag enabled. If the branch does not
match, the flag is defined in the CI cabal.project.local as being
explicitly disabled.